refactor(plot): improve directory structure in bokeh (#275)
* refactor(plot): move util.py into util directory


* refactor(plot): move legend.py into util directory


* refactor(plot): move color_selector.py into util directory


* chore: pass flake8


* refactor(plot): move common methods into util.py in Bokeh class


* chore: add copyright


* move hover-related classes to hover.py


---------
