+++ src/caret_analyze/plot/visualize_lib/bokeh/stacked_bar_source.py
@@ -19,7 +19,7 @@ from typing import Dict, List
 from bokeh.models import HoverTool
 from bokeh.plotting import ColumnDataSource
 
-from .legend import HoverCreator, HoverKeys, HoverSource, LegendManager
+from .util import HoverCreator, HoverKeys, HoverSource, LegendManager
 
 
 class StackedBarSource:
